typescript for the part of the Context that are independend to rules
